Move Enable to IPConfig
