+++ Aura Operating System/Aura_OS/System/Shell/cmdIntr/Network/IPConfig.cs
@@ -133,12 +133,12 @@ namespace Aura_OS.System.Shell.cmdIntr.Network
 
                     if (ip != null && subnet != null && gw != null)
                     {
-                        NetworkInit.Enable(nic, ip, subnet, gw);
+                        IPConfig.Enable(nic, ip, subnet, gw);
                         Console.WriteLine("Config OK!");
                     }
                     else if (ip != null && subnet != null)
                     {
-                        NetworkInit.Enable(nic, ip, subnet, ip);
+                        IPConfig.Enable(nic, ip, subnet, ip);
                         Console.WriteLine("Config OK!");
                     }
                     else
